requirements.txt
anthropic>=0.50.0
python-frontmatter>=1.1.0
